edited ft_memmove.c ft_strlcat.c ft_strncat.c

+++ ft_memmove.c
@@ -6,7 +6,7 @@
 /*   By: oevtushe <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/10/27 11:12:31 by oevtushe          #+#    #+#             */
-/*   Updated: 2017/11/07 12:54:28 by oevtushe         ###   ########.fr       */
+/*   Updated: 2017/11/07 12:58:01 by oevtushe         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,12 +14,12 @@
 
 void	*ft_memmove(void *dst, const void *src, size_t len)
 {
-	size_t	i;
+	int	i;
 
 	if (src - dst > 0)
 	{
 		i = 0;
-		while (i < len)
+		while ((unsigned int)i < len)
 		{
 			((unsigned char*)dst)[i] = ((unsigned char*)src)[i];
 			++i;
@@ -27,7 +27,7 @@ void	*ft_memmove(void *dst, const void *src, size_t len)
 	}
 	else
 	{
-		i = len;
+		i = (int)len;
 		while (--i >= 0)
 			((unsigned char*)dst)[i] = ((unsigned char*)src)[i];
 	}
